+++ src/settings/settings.cpp
@@ -128,7 +128,7 @@ char *settings_get_phone(uint8_t index)
     return phone;
 }
 
-uint8_t settings_find_phone_id(char *phone)
+uint8_t settings_find_phone_id(const char *phone)
 {
     uint8_t id = 255;
     for(uint8_t index = 0; index < settings_get_phones_list_length(); index++)
@@ -149,7 +149,7 @@ bool settings_delete_phone_by_id(uint8_t phone_index)
     if (phone_index < MAX_PHONES)
     {
         memset(global_settings.phone[phone_index], 0, 20);
-        for (uint8_t index = phone_index; index < settings_get_phones_list_length() - 1; index++)
+        for (uint8_t index = phone_index; index < settings_get_phones_list_length(); index++)
         {
             memcpy(global_settings.phone[index], global_settings.phone[index+1], 20);
         }
@@ -173,7 +173,7 @@ uint8_t settings_get_phones_list_length()
     return lenght;
 }
 
-void settings_set_local(char *local)
+void settings_set_local(const char *local)
 {
     strncpy(global_settings.local, local, 128);
 }